requirements.txt
Flask
flask-mysqldb
flask-jwt-extended
python-dotenv
APScheduler
mysql-connector-python